Add /createsales and /editsales route
