+++ src/pages/EditSalesEntry.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import NavBar from "../NavBar";
+
+const EditSalesEntry = () => {
+  return (
+    <>
+      <NavBar />
+      <h1>From edit</h1>
+    </>
+  );
+};
+
+export default EditSalesEntry;
